+++ src/const/const.js
@@ -1,0 +1,9 @@
+// optimization status codes
+// status codes for optimization direction 
+export const STATUS = {
+    OPTIMIZED: 0,
+    FAILED_OPT: 1,
+    NOT_OPT: -1,
+    MAX_UNDERFLOW: -4
+}
+
